separated out llvm worker classes for readability
