+++ public/codal-wasm/llvm-worker.js
@@ -1,7 +1,10 @@
 import FileSystem from "./modules/FileSystem.mjs";
 
 import LlvmBoxProcess from "./modules/LlvmBoxProcess.mjs";
-import ClangdProcess from "./modules/ClangdProcess.mjs"
+import ClangdProcess from "./modules/ClangdProcess.mjs";
+
+import { LSPUtil } from "./lsp.mjs";
+import { Clangd } from "./clangd-server.mjs";
 
 class LLVM {
     initialised = false;
@@ -66,8 +69,6 @@ class LLVM {
             '-o','../include/MicroBit.h.pch','-c', '/libraries/codal-microbit-v2/model/MicroBit.h'
         );
         
-        //temporary test to try to get clangd to work
-
         postMessage({
             target: "worker",
             type: "info",
@@ -101,9 +102,17 @@ class LLVM {
                 "character": 10
               }
         }) );
-        clangd.stdin.write(message2);
 
         llvm.run('clangd'); // '--log=verbose'
+
+        // await new Promise(resolve => setTimeout(function() {
+        //     console.log('Delayed task completed');
+        //     resolve();
+        //   }, 100));
+
+        // clangd.stdin.write(message2);     
+
+        console.log("After clangd runs")
     }
 
     onprocessstart = () => {};
@@ -263,9 +272,6 @@ function isError(stderr) {
 }
 
 async function clean() {
-    await llvm.clangdTest();
-
-
     let workingDir = await llvm.fileSystem.FS.analyzePath('/working/');
     let filesToRemove = workingDir.object.contents;
 
@@ -374,6 +380,7 @@ async function handleCompileRequest(files) {
         type: "compile-complete",
     })
 
+    await llvm.clangdTest();
     await clean();
 }
 
@@ -384,164 +391,3 @@ async function handleClangdRequest(request) {
 }
 
 const llvm = new LLVM();
-
-
-class LSPUtil {
-
-    HTTPWrapper = (json) => {
-        return `Content-Length: ${json.length}\r\n\r\n${json}`;
-    }
-
-    LSPCallWrapper = (method, params = {}) => {
-        const json = {
-            jsonrpc: "2.0",
-            id: 1,
-            method: method,
-            params: params,
-        }
-
-        return JSON.stringify(json);
-    }
-
-    LSPNotifyWrapper = (method, params = {}) => {
-        const json = {
-            jsonrpc: "2.0",
-            method: method,
-            params: params,
-        }
-
-        return JSON.stringify(json);
-    }
-}
-
-class StdStream {
-    buffer = [];
-
-    onPut = (char) => {
-        console.log(char)
-    }
-
-    write(data) {
-        for (let i=0; i<data.length; i++){
-            this.buffer.push(data[i]);
-        }
-    }
-
-    read() {
-        const bufStr = this.buffer.join('');
-        this.clear()
-        return bufStr;
-    }
-
-    get = () => {
-        if (this.buffer.length == 0) return null
-        const c = this.buffer.shift().charCodeAt(0);
-        return c;
-    }
-    
-    put = (byte) => {
-        const char = String.fromCharCode(byte);
-        this.buffer.push(char);
-        this.onPut(char);
-    }
-
-    clear() {
-        this.buffer = [];
-    }
-}
-
-class Clangd {
-    module;
-    stdin;
-    stdout;
-    stderr;
-
-    constructor (module) {
-        this.module = module;
-
-        this.initStreams();
-    }
-
-    initStreams() {
-        const stdin = new StdStream();
-
-        const stdout = new StdStream();
-        stdout.onPut = (char) => { this.stdoutOnPutHandler(char) };
-        
-        const stderr = new StdStream();
-        stderr.onPut = (char) => {
-            if (char === "\n") {
-                const line = stderr.read();
-                if (line.startsWith("E")) console.error(line);
-                else console.log(line);
-            }
-        }
-
-        this.module.FS.init(stdin.get, stdout.put, stderr.put);
-
-        this.stdin = stdin;
-        this.stdout = stdout;
-        this.stderr = stderr;
-    }
-
-    //Handler for stdout. Processes LSP response messages as per the specification:
-    //https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#baseProtocol
-
-    _separator_ = ": ";
-    _terminator_ = "\r\n";
-    _fullTerminator = `${this._terminator_}${this._terminator_}`
-
-    responseHeaders = {};
-    contentReady = false;
-    outBuffer = [];
-    outBufferLen = 4;
-
-    contentRemaining = 0;
-
-    
-    stdoutOnPutHandler(char) {
-        if (this.contentReady) {
-            this.contentRemaining--;
-            
-            if (this.contentRemaining == 0) {
-                this.contentReady = false;
-                const content = this.stdout.read();
-                const response = JSON.parse(content);
-                
-                postMessage({
-                    target: "clangd",
-                    type: "response",
-                    body: response
-                })
-            }
-            return;
-        }
-
-        //Push to a small buffer of chars to check the most recent characters read
-        if (this.outBuffer.length >= this.outBufferLen) this.outBuffer.shift();
-        this.outBuffer.push(char);
-
-        if(this.outBuffer.join('') === this._fullTerminator) {
-            // console.log("All headers received. Reading content");
-
-            this.stdout.clear()  //there shouldn't be any useful data left in the buffer
-
-            //Prepare to read content
-            if (!this.responseHeaders["Content-Length"]) console.error("Expected header 'Content-Length' was not found! Cannot continue reading LSP response");
-            else {
-                this.contentReady = true;
-                this.contentRemaining = this.responseHeaders["Content-Length"];
-            }
-        }
-        else if(this.outBuffer.slice(-2).join('') === this._terminator_) {
-            this.stdout.buffer.pop();                   //pop \r\n separator
-            this.stdout.buffer.pop();
-
-            const header = this.stdout.read();
-            // console.log(`Received Header: '${header}'`);
-            
-            const kvp = header.split(this._separator_);
-            this.responseHeaders[kvp[0]] = kvp[1];
-        }
-    }
-}
