+++ public/index.html
@@ -2,10 +2,10 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
+    <link rel="icon" href="%PUBLIC_URL%/favicon_cpp.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#6c4bc1" />
-    <title>micro:bit Python Editor</title>
+    <title>micro:bit C++ Editor</title>
     <meta property="og:title" content="micro:bit Python Editor" />
     <meta
       property="og:description"
